Finished the Handle implementation
The most major change is the round changing logic. Previously it was invoked by either the controller or the Heartbeat timer. Now it's unified under the applicable Handle class. This makes it possible to nicely handle the different cases in the Game logic, while offloading the actual state determinations to the handles. I also removed the whole reset functionality: it was impossible to get working without a major change to the architecture. Also, sorry for the insane amount of synchronized
methods: it took quite a bit of debugging to get here, I'll review and remove them later as necessary.