More accessible /dev/null like even sink
In cases where we use a 1:1 model for subscribers / listeners (e.g it doesn't make too much sense for a Handle to send events to multiple games) it is much easier to use a global sink that ignores any events sent to it than null-checking everywhere. Since this might come up later it only makes sense to have it somewhere reusable.