Improved build scripts
Added multiprocessing to the file list generation and svg conversion. The latter offered a 5x improvement in a build, where all the svg files had to be regenerated. Even the former is a big boost (3x), but it's much less noticable (down to ~0.1s from ~0.3 on my machine). Also using a lot more list comprehensions.